#237 - explore another major
#237 explore another major

From Student Profile, look at the remaining courses. Click Da Button
