+++ student-info-system/schooladmin/forms.py
@@ -8,46 +8,6 @@ from sis.models import (Course, CoursePrerequisite, Major, Professor, Section, S
 from sis.forms.utils import *
 
 
-class MajorCreationForm(forms.ModelForm):
-    abbreviation = UpperFormField(max_length=6, help_text='Abbreviation')
-    title = forms.CharField(max_length=256)
-    description = forms.CharField(max_length=256,
-                                  required=False,
-                                  widget=forms.Textarea(attrs={'rows': 3}))
-    contact = forms.ModelChoiceField(queryset=Profile.staff())
-
-    class Meta:
-        model = Major
-        fields = ('abbreviation', 'title', 'description', 'contact')
-
-
-class MajorEditForm(forms.ModelForm):
-    abbreviation = UpperFormField(max_length=6, help_text='Abbreviation')
-    title = forms.CharField(max_length=256)
-    description = forms.CharField(max_length=256,
-                                  required=False,
-                                  widget=forms.Textarea(attrs={'rows': 3}))
-    contact = forms.ModelChoiceField(queryset=Profile.staff())
-    courses_required = CourseChoiceField(queryset=Course.objects.all().order_by(
-        'major', 'catalog_number'),
-                                         widget=forms.CheckboxSelectMultiple,
-                                         required=False)
-
-    class Meta:
-        model = Major
-        fields = ('abbreviation', 'title', 'description', 'contact', 'courses_required')
-
-    def __init__(self, *args, **kwargs):
-        super(MajorEditForm, self).__init__(*args, **kwargs)
-        # we defer loading of professors until we know what major is chosen
-        if kwargs['instance']:
-            the_major = kwargs['instance']
-            if the_major:
-                self.fields['contact'].queryset = Profile.objects.filter(
-                    Q(role=Profile.ACCESS_ADMIN) |
-                    (Q(role=Profile.ACCESS_PROFESSOR) & Q(professor__major=the_major)))
-
-
 class CourseCreationForm(forms.ModelForm):
     major = forms.ModelChoiceField(queryset=Major.objects.all())
     major.widget.attrs.update({'class': 'major_sel selectpicker'})
