+++ student-info-system/sis/tables/courses.py
@@ -1,6 +1,6 @@
 import django_tables2 as tables
 
-from sis.models import Course
+from sis.models import Course, SectionStudent
 from sis.tables import *
 
 
@@ -39,10 +39,16 @@ class MajorCoursesMetTable(tables.Table):
     title = ClassyColumn(css_class_base='coursetitle')
     credits_earned = ClassyColumn(css_class_base='credits')
     met = ClassyBooleanColumn(css_class_base='met')
+    letter_grade = ClassyColumn(verbose_name="Grade",
+                                accessor='grade',
+                                css_class_base='lettergrade',
+                                order_by=('grade'))
+    def render_letter_grade(self, value):
+        return SectionStudent.letter_grade_for(value)
 
     class Meta:
         model = Course
         template_name = "django_tables2/bootstrap.html"
-        fields = ('major', 'catalog_number', 'title', 'credits_earned', 'met')
+        fields = ('major', 'catalog_number', 'title', 'credits_earned', 'letter_grade', 'met')
         row_attrs = {'class': 'course_row', 'data-id': lambda record: record.pk}
         attrs = {"class": 'majorcourse_table'}
